Separating classes to different files to make reading easier and updating the usage documentation to remove reference to decoration (old and redundant)
